using channels for saving metrics. ISSUE: segfaults in attribute_activity

+++ src/tool/hpcrun/gpu/intel/opencl-api.h
@@ -17,5 +17,5 @@ typedef struct profilingData
 } profilingData;
 #endif
 
-void opencl_subscriber_callback(opencl_call, uint32_t *);
+void opencl_subscriber_callback(opencl_call, uint64_t *);
 void opencl_buffer_completion_callback(cl_event, cl_int, void *);
